requirements.txt
## These requirements were autogenerated by pipenv# To regenerate from the project's Pipfile, run:##    pipenv lock --requirements#-i https://pypi.org/simpleastroid==2.14.1; python_full_version >= '3.7.2'attrs==22.2.0; python_version >= '3.6'automat==22.10.0certifi==2022.12.7; python_version >= '3.6'cffi==1.15.1charset-normalizer==3.0.1colorama==0.4.6; sys_platform == 'win32'constantly==15.1.0cryptography==39.0.1; python_version >= '3.6'cssselect==1.2.0; python_version >= '3.7'dill==0.3.6; python_version < '3.11'dnspython==2.3.0; python_version >= '3.7' and python_version < '4.0'fake-useragent==1.1.1faker==16.8.1; python_version >= '3.7'filelock==3.9.0; python_version >= '3.7'hyperlink==21.0.0idna==3.4; python_version >= '3.5'importlib-metadata==4.13.0; python_version < '3.8'importlib-resources==5.10.2; python_version < '3.10'incremental==22.10.0isort==5.11.5; python_version >= '3.7'itemadapter==0.7.0; python_version >= '3.6'itemloaders==1.0.6; python_version >= '3.6'jmespath==1.0.1; python_version >= '3.7'lazy-object-proxy==1.9.0; python_version >= '3.7'lxml==4.9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'mccabe==0.7.0; python_version >= '3.6'packaging==23.0; python_version >= '3.7'parsel==1.7.0; python_version >= '3.7'platformdirs==3.0.0; python_version >= '3.7'protego==0.2.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'pyasn1-modules==0.2.8pyasn1==0.4.8pycparser==2.21pydispatcher==2.0.6; platform_python_implementation == 'CPython'pylint==2.16.1pymongo==4.3.3pyopenssl==23.0.0; python_version >= '3.6'python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'queuelib==1.6.2; python_version >= '3.5'requests-file==1.5.1requests==2.28.2; python_version >= '3.7' and python_version < '4.0'scrapeops-scrapy-proxy-sdk==1.0scrapy-fake-useragent==1.4.4scrapy==2.8.0service-identity==21.1.0six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'tldextract==3.4.0; python_version >= '3.7'tomli==2.0.1; python_version < '3.11'tomlkit==0.11.6; python_version >= '3.6'twisted-iocpsupport==1.0.2; platform_system == 'Windows'twisted==22.10.0; python_full_version >= '3.7.1'typed-ast==1.5.4; python_version < '3.8' and implementation_name == 'cpython'typing-extensions==4.4.0; python_version < '3.11'urllib3==1.26.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'w3lib==2.1.1; python_version >= '3.7'wrapt==1.14.1; python_version < '3.11'zipp==3.13.0; python_version < '3.10'pycryptodomezope.interface==5.5.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'